All parsing working except null string

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mvalient <mvalient@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/12 12:00:14 by mvalient          #+#    #+#             */
-/*   Updated: 2022/09/15 14:10:38 by mvalient         ###   ########.fr       */
+/*   Updated: 2022/09/15 17:28:54 by mvalient         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@
 # include "libft/libft.h"
 
 int	ft_printf(char const *s, ...);
-int	ft_putnbr_base(int nbr, char *base);
+int	ft_putnbr_base(unsigned long int nbr, char *base);
 
 #endif
